+++ src/include/cell_move_router/RegionCalculator/WeightedOptimalRegion.hpp
@@ -3,9 +3,9 @@ namespace cell_move_router {
 namespace RegionCalculator {
 
 class WeightedOptimalRegion : RegionCalculator {
-  int weightElement(std::vector<std::pair<int, double>> &V, const int L,
-                    const int R, double W);
-  std::pair<int, int> getWeightedMedium(std::vector<std::pair<int, double>> &V);
+  int weightElement(std::vector<std::pair<int, long long>> &V, const int L,
+                    const int R, long long W);
+  std::pair<int, int> getWeightedMedium(std::vector<std::pair<int, long long>> &V);
 
 public:
   WeightedOptimalRegion() {}
